feat: chatdb template
